allow_missing_fields defaults to false
My intent was to allow things like Log messages to map if defined via mapping_rules yaml.  This didn't work with Log because the defines for the different levels changed, so rqt's console display would go crazy about invalid severity.  If this is useful, it will have to be added to the yaml rules definition to be exposed.

Signed-off-by:  Douglas Smith <douglas.smith@swri.org>